backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
schedule==1.2.0
prometheus-client==0.19.0
